multiagent-collab-cs-call-center-analysis/requirements-crewai.txt
crewai[tools]>=0.86.0
PyYAML>=6.0.1
